requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.8.2
pydantic-settings==2.4.0
python-multipart==0.0.9
PyYAML==6.0.2
pandas==2.2.2
numpy==2.0.1
networkx==3.3

streamlit==1.37.1
pytest==8.3.2
httpx==0.27.0
